DTO and password encoder

+++ backend/src/main/java/com/study/backend/service/ProductService.java
@@ -1,13 +1,15 @@
 package com.study.backend.service;
 
 import com.study.backend.entity.Product;
-import com.study.backend.entity.User;
+import com.study.backend.user.User;
+import com.study.backend.exception.NotFoundException;
 import com.study.backend.repository.ProductRepository;
 import com.study.backend.repository.UserRepository;
+import jakarta.persistence.EntityNotFoundException;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import javax.persistence.EntityNotFoundException;
 import java.util.List;
 
 @Service
@@ -24,9 +26,17 @@ public class ProductService {
         return productRepository.save(product);
     }
 
-    public void updateProduct(Product product) {
+    public Product updateProduct(Long id, @Valid Product product) {
+        Product existingProduct = getProductById(id);
+        if (existingProduct == null) {
+            throw new NotFoundException("Product with id: " + id + " not existing");
+        }
         // Perform any necessary validation
-        productRepository.save(product);
+        existingProduct.setName(product.getName());
+        existingProduct.setPrice(product.getPrice());
+        // Perform any necessary validation
+        productRepository.save(existingProduct);
+        return existingProduct;
     }
 
     public List<Product> getAllProduct() {
@@ -38,6 +48,10 @@ public class ProductService {
     }
 
     public void deleteProductById(Long id) {
+        Product existingProduct = getProductById(id);
+        if (existingProduct == null) {
+            throw new NotFoundException("Product with id: " + id + " not existing");
+        }
         productRepository.deleteById(id);
     }
